requirements.txt
anyio==3.7.0
certifi==2023.5.7
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==20.3
sniffio==1.3.0
